video listing updated
