+++ src/frontend/components/modal/modal.css
@@ -1,0 +1,37 @@
+@import '../../utility/styles.css';
+
+/* Modal */
+.modal__background {
+  opacity: 0.9;
+}
+.modal__content {
+  width: 100%;
+  max-width: 25rem;
+  border: none;
+  background-color: var(--dismiss);
+  color: var(--white);
+  padding: 1.25rem 1.5rem;
+  font-size: 1.25rem;
+}
+.playlist__option {
+  display: flex;
+  align-items: center;
+  margin: 1rem 0;
+}
+.playlist__input {
+  width: 100%;
+  margin: 1rem 0 0 0;
+  display: block;
+  padding: 0.5rem;
+  font-size: 1rem;
+  background-color: var(--dismiss);
+  color: var(--white);
+  border-radius: 5px;
+  border: 1px solid var(--white);
+  outline: none;
+}
+.playlist__error {
+  font-size: 1rem;
+  color: var(--tag);
+  margin: 0.5rem 0.5rem 1.25rem;
+}
