+++ src/frontend/pages/videos/videos.css
@@ -1,1 +1,16 @@
 @import '../../utility/styles.css';
+
+.videogrid {
+  display: grid;
+  grid-template-columns: 100%;
+  grid-template-areas:
+    'main'
+    'sidebar';
+}
+
+@media screen and (min-width: 1024px) {
+  .videogrid {
+    grid-template-columns: 20% 80%;
+    grid-template-areas: 'sidebar main';
+  }
+}
